# internal/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"encoding/json"
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type FinalResult struct {
	ID        int64     `json:"id"`
	GameID    int64     `json:"game_id"`
	TeamID    int64     `json:"team_id"`
	Score     int32     `json:"score"`
	Position  *int32    `json:"position"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Game struct {
	ID                   int64              `json:"id"`
	Name                 *string            `json:"name"`
	Organizer            *string            `json:"organizer"`
	StartsAt             pgtype.Timestamptz `json:"starts_at"`
	EndsAt               pgtype.Timestamptz `json:"ends_at"`
	CreatedAt            time.Time          `json:"created_at"`
	UpdatedAt            time.Time          `json:"updated_at"`
	AvatarUrl            *string            `json:"avatar_url"`
	SiteUrl              *string            `json:"site_url"`
	CtftimeUrl           *string            `json:"ctftime_url"`
	Finalized            bool               `json:"finalized"`
	FinalizedAt          pgtype.Timestamptz `json:"finalized_at"`
	RegistrationOpensAt  pgtype.Timestamptz `json:"registration_opens_at"`
	RegistrationClosesAt pgtype.Timestamptz `json:"registration_closes_at"`
	ScoreboardOpensAt    pgtype.Timestamptz `json:"scoreboard_opens_at"`
	ScoreboardClosesAt   pgtype.Timestamptz `json:"scoreboard_closes_at"`
	VpnUrl               *string            `json:"vpn_url"`
	VpnConfigUrl         *string            `json:"vpn_config_url"`
	AccessInstructions   *string            `json:"access_instructions"`
	AccessSecret         *string            `json:"access_secret"`
}

type GameTeam struct {
	ID              int64           `json:"id"`
	GameID          int64           `json:"game_id"`
	TeamID          int64           `json:"team_id"`
	IpAddress       *string         `json:"ip_address"`
	Ctf01dID        *string         `json:"ctf01d_id"`
	Ctf01dOverrides json.RawMessage `json:"ctf01d_overrides"`
	TeamType        *string         `json:"team_type"`
	Order           int32           `json:"order"`
	CreatedAt       time.Time       `json:"created_at"`
	UpdatedAt       time.Time       `json:"updated_at"`
}

type GamesService struct {
	GameID    int64 `json:"game_id"`
	ServiceID int64 `json:"service_id"`
}

type Result struct {
	ID        int64     `json:"id"`
	GameID    int64     `json:"game_id"`
	TeamID    int64     `json:"team_id"`
	Score     *int32    `json:"score"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Service struct {
	ID                  int64              `json:"id"`
	Name                string             `json:"name"`
	PublicDescription   *string            `json:"public_description"`
	PrivateDescription  *string            `json:"private_description"`
	Author              *string            `json:"author"`
	Copyright           *string            `json:"copyright"`
	AvatarUrl           *string            `json:"avatar_url"`
	Public              bool               `json:"public"`
	CreatedAt           time.Time          `json:"created_at"`
	UpdatedAt           time.Time          `json:"updated_at"`
	ServiceArchiveUrl   *string            `json:"service_archive_url"`
	CheckerArchiveUrl   *string            `json:"checker_archive_url"`
	WriteupUrl          *string            `json:"writeup_url"`
	ExploitsUrl         *string            `json:"exploits_url"`
	CheckStatus         string             `json:"check_status"`
	CheckedAt           pgtype.Timestamptz `json:"checked_at"`
	ServiceLocalPath    *string            `json:"service_local_path"`
	ServiceLocalSize    *int32             `json:"service_local_size"`
	ServiceLocalSha256  *string            `json:"service_local_sha256"`
	ServiceDownloadedAt pgtype.Timestamptz `json:"service_downloaded_at"`
	CheckerLocalPath    *string            `json:"checker_local_path"`
	CheckerLocalSize    *int32             `json:"checker_local_size"`
	CheckerLocalSha256  *string            `json:"checker_local_sha256"`
	CheckerDownloadedAt pgtype.Timestamptz `json:"checker_downloaded_at"`
	Ctf01dTraining      json.RawMessage    `json:"ctf01d_training"`
}

type Team struct {
	ID           int64     `json:"id"`
	Name         string    `json:"name"`
	Description  *string   `json:"description"`
	Website      *string   `json:"website"`
	AvatarUrl    *string   `json:"avatar_url"`
	CaptainID    *int32    `json:"captain_id"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
	UniversityID *int64    `json:"university_id"`
}

type TeamMembership struct {
	ID        int64     `json:"id"`
	TeamID    int64     `json:"team_id"`
	UserID    int64     `json:"user_id"`
	Role      *string   `json:"role"`
	Status    *string   `json:"status"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type TeamMembershipEvent struct {
	ID         int64     `json:"id"`
	TeamID     int64     `json:"team_id"`
	UserID     int64     `json:"user_id"`
	ActorID    *int32    `json:"actor_id"`
	Action     string    `json:"action"`
	FromRole   *string   `json:"from_role"`
	ToRole     *string   `json:"to_role"`
	FromStatus *string   `json:"from_status"`
	ToStatus   *string   `json:"to_status"`
	CreatedAt  time.Time `json:"created_at"`
	UpdatedAt  time.Time `json:"updated_at"`
}

type University struct {
	ID        int64     `json:"id"`
	Name      *string   `json:"name"`
	SiteUrl   *string   `json:"site_url"`
	AvatarUrl *string   `json:"avatar_url"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type User struct {
	ID             int64              `json:"id"`
	UserName       string             `json:"user_name"`
	DisplayName    string             `json:"display_name"`
	Role           string             `json:"role"`
	Rating         int32              `json:"rating"`
	AvatarUrl      *string            `json:"avatar_url"`
	PasswordDigest *string            `json:"password_digest"`
	CreatedAt      time.Time          `json:"created_at"`
	UpdatedAt      time.Time          `json:"updated_at"`
	Bio            *string            `json:"bio"`
	Telegram       *string            `json:"telegram"`
	Github         *string            `json:"github"`
	Email          *string            `json:"email"`
	IsBlocked      bool               `json:"is_blocked"`
	BlockedAt      pgtype.Timestamptz `json:"blocked_at"`
	LastLoginIp    *string            `json:"last_login_ip"`
	LastLoginAt    pgtype.Timestamptz `json:"last_login_at"`
}

type UserSession struct {
	ID         int64              `json:"id"`
	UserID     int64              `json:"user_id"`
	Jti        string             `json:"jti"`
	IpAddress  *string            `json:"ip_address"`
	UserAgent  *string            `json:"user_agent"`
	CreatedAt  time.Time          `json:"created_at"`
	LastSeenAt time.Time          `json:"last_seen_at"`
	ExpiresAt  time.Time          `json:"expires_at"`
	RevokedAt  pgtype.Timestamptz `json:"revoked_at"`
}

type Writeup struct {
	ID        int64     `json:"id"`
	GameID    int64     `json:"game_id"`
	TeamID    int64     `json:"team_id"`
	Title     string    `json:"title"`
	Url       string    `json:"url"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}